TASK: Remove master default of the clone preset command
With the default value and differing preset could only be added like this ``clone:preset --preset-name __your_name__``.
Before the preset name could be passed to the controller immediately ``clone:preset __your_name__`` we consider
this to be the most useful interface for this command so we revert the defaulr=master change.